enable auto-discovery of detectors
