+++ frigate/detectors/detection_api.py
@@ -1,15 +1,15 @@
 import logging
-
 from abc import ABC, abstractmethod
-from typing import Dict
 
 
 logger = logging.getLogger(__name__)
 
 
 class DetectionApi(ABC):
+    type_key: str
+
     @abstractmethod
-    def __init__(self, det_device=None, model_config=None):
+    def __init__(self, detector_config):
         pass
 
     @abstractmethod
