+++ frigate/object_detection.py
@@ -11,7 +11,7 @@ import numpy as np
 from setproctitle import setproctitle
 
 from frigate.enums import InputTensorEnum
-from frigate.detectors import create_detector, DetectorTypeEnum
+from frigate.detectors import create_detector
 
 from frigate.util import EventsPerSecond, SharedMemoryFrameManager, listen, load_labels
